+++ R/reexports.R
@@ -1,0 +1,33 @@
+#' Pipe operator
+#'
+#' See \code{magrittr::\link[magrittr]{\%>\%}} for details.
+#'
+#' @name %>%
+#' @rdname pipe
+#' @keywords internal
+#' @export
+#' @importFrom magrittr %>%
+#' @usage lhs \%>\% rhs
+NULL
+
+
+#' @importFrom fablelite features features_at features_if features_all
+#' @export
+fablelite::features
+
+#' @export
+fablelite::features_at
+
+#' @export
+fablelite::features_if
+
+#' @export
+fablelite::features_all
+
+#' @importFrom tsibble as_tsibble n_keys
+#' @export
+tsibble::as_tsibble
+
+#' @export
+tsibble::n_keys
+